Backend_Pull_Request

+++ app/backend/src/controllers/otp.controller.js
@@ -6,9 +6,7 @@ import asyncHandler from "../utils/asyncHandler.js";
 import User from "../models/user.model.js";
 import OTP from "../models/userOTP.model.js";
 import mobileOTP from "../models/mobileOTP.model.js";
-// import { generateOTPToken } from "../utils/auth.js";
-import bcrypt from "bcrypt";
-
+import { generateOTPToken } from "../utils/auth.js";
 import { verifyHashedData } from "../utils/hashData.js";
 
 export const verifyOTP = asyncHandler(async (req, res) => {
@@ -21,9 +19,8 @@ export const verifyOTP = asyncHandler(async (req, res) => {
     const userOTP = await OTP.findOne({
       email: decoded.email || decoded.userData.email,
     });
-    console.log("opt=",otp,"userOTP= ", userOTP.otp)
-    const match =  await bcrypt.compare(otp, userOTP.otp);
-    console.log(match);
+
+    const match = await verifyHashedData(otp, userOTP.otp);
     if (!match) {
       throw new ApiError(400, "Invalid OTP.");
     }
@@ -40,15 +37,15 @@ export const verifyOTP = asyncHandler(async (req, res) => {
       firstName: decoded.firstName || decoded.userData.firstName,
       lastName: decoded.lastName || decoded.userData.lastName,
       password: decoded.password || decoded.userData.password,
-      // contact: decoded.contact || decoded.userData.contact|| null,
-      // avatar: decoded.avatar || decoded.userData.avatar,
+      contact: decoded.contact || decoded.userData.contact,
+      avatar: decoded.avatar || decoded.userData.avatar,
       isVerified: true, // Mark the user as verified
     });
 
-    // const incomingAvatar = await uploadOnCloudinary(user.avatar);
-    // if (!incomingAvatar) {
-    //   throw new ApiError(400, "Avatar upload failed.");
-    // }
+    const incomingAvatar = await uploadOnCloudinary(user.avatar);
+    if (!incomingAvatar) {
+      throw new ApiError(400, "Avatar upload failed.");
+    }
     await user.save();
     await OTP.deleteMany({ email: user.email });
     return res
@@ -78,24 +75,23 @@ export const resendOTPCode = asyncHandler(async (req, res) => {
     });
   }
 
-  // const userData = {
-  //   email: decoded.email,
-  //   firstName: decoded.firstName,
-  //   lastName: decoded.lastName,
-  //   password: decoded.password,
-  //   contact: decoded.contact,
-  //   // avatar: decoded.avatar,
-  // };
+  const userData = {
+    email: decoded.email,
+    firstName: decoded.firstName,
+    lastName: decoded.lastName,
+    password: decoded.password,
+    contact: decoded.contact,
+    avatar: decoded.avatar,
+  };
 
-  // const newData = generateOTPToken({ userData });
+  const newData = generateOTPToken({ userData });
   await sendEmail({ email });
   res.status(200).json({
     status: "PENDING",
     message: "OTP resent successfully. Please check your email.",
-    token: token,
+    token: newData.token,
   });
 });
-
 
 export const verifyMobileOTP = asyncHandler(async (req, res) => {
   const { token, otp } = req.body;
@@ -120,18 +116,18 @@ export const verifyMobileOTP = asyncHandler(async (req, res) => {
     throw new ApiError(409, "User with this contact already exists.");
   }
   const user = new User({
-    // email: decoded.email,
+    email: decoded.email,
     firstName: decoded.firstName,
     lastName: decoded.lastName,
     password: decoded.password,
     contact: decoded.contact,
-    // avatar: decoded.avatar,
+    avatar: decoded.avatar,
     isVerified: true,
   });
-  // const incomingAvatar = await uploadOnCloudinary(user.avatar);
-  // if (!incomingAvatar) {
-  //   throw new ApiError(400, "Avatar upload failed.");
-  // }
+  const incomingAvatar = await uploadOnCloudinary(user.avatar);
+  if (!incomingAvatar) {
+    throw new ApiError(400, "Avatar upload failed.");
+  }
   await user.save();
 
   const deletedOTP = await mobileOTP.deleteMany({ contact: decoded.contact });
@@ -159,19 +155,19 @@ export const resendMobileOTP = asyncHandler(async (req, res) => {
       message: "User is already verified. No need to resend OTP.",
     });
   }
-  // const userData = {
-  //   email: decoded.email,
-  //   firstName: decoded.firstName,
-  //   lastName: decoded.lastName,
-  //   password: decoded.password,
-  //   contact: decoded.contact,
-  //   // avatar: decoded.avatar,
-  // };
-  // const newData = generateOTPToken({ userData });
+  const userData = {
+    email: decoded.email,
+    firstName: decoded.firstName,
+    lastName: decoded.lastName,
+    password: decoded.password,
+    contact: decoded.contact,
+    avatar: decoded.avatar,
+  };
+  const newData = generateOTPToken({ userData });
   await sendMessage(decoded.contact);
   res.status(200).json({
     status: "PENDING",
     message: "OTP resent successfully. Please check your phone.",
-    token: token,
+    token: newData.token,
   });
 });
